Skills Overhaul and testing
NOTES:

The skills system has been overhauled and reworked so that there are now only three attributes:
- Body
- Mind
- Soul

It was agreedt that this simpler approach would be more in line with the MC play style.

In addition, The skills stystem has been revamped, there are now less skills and a more flexible and cleaner (the mess in the skilll collection has been cleaned up) way of adding and initilising skills. To acompany this, a roll table and "dice" system has been created.

TODO:

- The skill Forestry has been tested but the event handler is a little messy.
- As several skills have the same effects once the roll has been determined, this shall be broken into a seperate skillEffect class for comonality